Add environment sagas and reducer
