+++ ui/src/ducks/reducer.js
@@ -9,6 +9,7 @@ import layout from './app/layout';
 import notifications from './app/notifications';
 import salt from './app/salt';
 import monitoring from './app/monitoring';
+import solutions from './app/solutions';
 
 const rootReducer = combineReducers({
   config,
@@ -20,8 +21,9 @@ const rootReducer = combineReducers({
     notifications,
     salt,
     monitoring,
-    volumes
-  })
+    volumes,
+    solutions,
+  }),
 });
 
 export default rootReducer;
